+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add support for ``SlepcInitialize(filename,help,ierr)``, ``SlepcInitialize(ierr)`` in
   addition to current ``SlepcInitialize(filename,ierr)``
 - New `DSGSVD` type for the generalized SVD.
+- New contour integral method in `DSNEP`.
 
 ### Changed